pyproject.toml
[project]
name = "unubot"
version = "0.1.0"
description = "Community support bot for the UNU Community Discord — unu scooter and LibreScoot help"
requires-python = ">=3.11"
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "discord.py>=2.4.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
unubot = "unubot.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
